Extract one unified FakeRiderScoreDao and reuse across all tests
