+++ app/src/test/java/net/yakavenka/trialsscore/data/ScoreSummaryRepositoryTest.kt
@@ -10,7 +10,7 @@ import org.hamcrest.Matchers.not
 import org.junit.Test
 
 class ScoreSummaryRepositoryTest {
-    private val dao: RiderScoreDaoFake = RiderScoreDaoFake()
+    private val dao: FakeRiderScoreDao = FakeRiderScoreDao()
 
     private val sut = ScoreSummaryRepository(dao)
 
@@ -22,52 +22,4 @@ class ScoreSummaryRepositoryTest {
         assertThat(actual, not(empty()))
     }
 
-    class RiderScoreDaoFake : RiderScoreDao {
-        val summary = mutableListOf<RiderScoreSummary>()
-
-        override fun getAll(): Flow<List<RiderScoreAggregate>> {
-            TODO("Not yet implemented")
-        }
-
-        override fun getRider(riderId: Int): Flow<RiderScore> {
-            TODO("Not yet implemented")
-        }
-
-        override fun sectionScores(riderId: Int, loopNumber: Int): Flow<List<SectionScore>> {
-            TODO("Not yet implemented")
-        }
-
-        override suspend fun updateSectionScore(sectionScore: SectionScore) {
-            TODO("Not yet implemented")
-        }
-
-        override suspend fun deleteAllRiders() {
-            TODO("Not yet implemented")
-        }
-
-        override suspend fun deleteAllScores() {
-            TODO("Not yet implemented")
-        }
-
-        override suspend fun deleteRiderScores(riderId: Int) {
-            TODO("Not yet implemented")
-        }
-
-        override suspend fun insertAll(sectionScores: List<SectionScore>) {
-            TODO("Not yet implemented")
-        }
-
-        override suspend fun addRider(riderScore: RiderScore) {
-            TODO("Not yet implemented")
-        }
-
-        override fun fetchSummary(): Flow<List<RiderScoreSummary>> = flow {
-            emit(summary)
-        }
-
-        override suspend fun updateRider(riderScore: RiderScore) {
-            TODO("Not yet implemented")
-        }
-
-    }
 }
